+++ tomcat/src/main/java/org/apache/coyote/http11/message/request/requestline/Method.java
@@ -1,0 +1,8 @@
+package org.apache.coyote.http11.message.request.requestline;
+
+public enum Method {
+
+    GET,
+    POST,
+    ;
+}
