+++ tomcat/src/main/java/org/apache/coyote/http11/message/request/requestline/RequestLine.java
@@ -1,7 +1,7 @@
-package org.apache.coyote.http11.request.requestline;
+package org.apache.coyote.http11.message.request.requestline;
 
-import org.apache.coyote.http11.Regex;
-import org.apache.coyote.http11.request.QueryParams;
+import org.apache.coyote.http11.message.Regex;
+import org.apache.coyote.http11.message.request.QueryParams;
 
 public class RequestLine {
