+++ tomcat/src/main/java/org/apache/coyote/http11/message/response/Response.java
@@ -1,14 +1,14 @@
-package org.apache.coyote.http11.response;
+package org.apache.coyote.http11.message.response;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
 import java.util.Map;
 import nextstep.jwp.util.ResourceLoader;
-import org.apache.coyote.http11.Regex;
-import org.apache.coyote.http11.header.Header;
-import org.apache.coyote.http11.response.header.ContentType;
-import org.apache.coyote.http11.response.header.Headers;
-import org.apache.coyote.http11.response.header.StatusCode;
+import org.apache.coyote.http11.message.Regex;
+import org.apache.coyote.http11.message.header.Header;
+import org.apache.coyote.http11.message.response.header.ContentType;
+import org.apache.coyote.http11.message.response.header.Headers;
+import org.apache.coyote.http11.message.response.header.StatusCode;
 
 public class Response {
