--- tomcat/src/main/java/org/apache/coyote/http11/request/requestline/Method.java
@@ -1,8 +1,0 @@
-package org.apache.coyote.http11.request.requestline;
-
-public enum Method {
-
-    GET,
-    POST,
-    ;
-}
